Compgraph demo
