+++ src/node_operations.ts
@@ -1,4 +1,3 @@
-import {Tensor} from "./base/Tensor.ts";
 import Node from "./Node.ts";
 import * as ops from "./base/tensor_operations.ts";
 
@@ -14,23 +13,30 @@ export interface BidirectionalOperation {
     bw: BackwardFunc;
 }
 
+// todo: what about scalar addition?
+// Pairwise tensor addition
 export const add: BidirectionalOperation = {
     name: "add",
 
     fw(parents: Node[], self: Node) {
+        //      src 1               src 2             dest
         ops.add(parents[0].primal, parents[1].primal, self.primal);
     },
 
     bw(parents: Node[], self: Node) {
-        // todo: validate if this works as expected (because of the changes to the tensor_ops + Tensor class)
+        //      src 1            src 2      dest
         ops.add(parents[0].grad, self.grad, parents[0].grad); // in-place op
         ops.add(parents[1].grad, self.grad, parents[1].grad); // in-place op
-        // parents[0].grad.add(self.grad, true);
-        // parents[1].grad.add(self.grad, true);
     }
 };
 
+// Does nothing.
 export const nop: BidirectionalOperation = {
+    name: "nop",
+
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     fw(parents: Node[], primal: Node) {},
+
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     bw(parents: Node[], primal: Node) {}
 };
